add frontend and multiple client/server debugs

+++ src/utils/cloudinary.js
@@ -1,32 +1,41 @@
 import dotenv from "dotenv";
 dotenv.config({
-    path : './.env'
+    path: './.env'
 })
-import {v2 as cloudinary} from "cloudinary"
+import { v2 as cloudinary } from "cloudinary"
 import fs from "fs"
+import { ApiError } from "../utils/ApiError.js";
 
 cloudinary.config({
-    cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-    api_key: process.env.CLOUDINARY_API_KEY , 
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
     api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
-    try{
-        if(!localFilePath) return null;
+    try {
+        if (!localFilePath) return null;
 
         const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
-        })
+        });
 
-        fs.unlinkSync(localFilePath);
-        return response;
+        if (fs.existsSync(localFilePath)) {
+            try {
+                fs.unlinkSync(localFilePath);
+            } catch (err) {
+                console.warn("🧹 File deletion error:", err.message);
+            }
+        }
+        return response.secure_url;
     }
-    catch(error){
+    catch (error) {
         console.error("Cloudinary Upload Error:", error)
-        fs.unlinkSync(localFilePath);
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
         return null;
     }
 }
 
-export {uploadOnCloudinary}
+export { uploadOnCloudinary }
